+++ docs/src/pages/customization/overrides.md
@@ -1,6 +1,6 @@
 # Overrides
 
-As components evolve in different contexts, Material-UI supports four different types of customization needs going from 
+As components evolve in different contexts, Material-UI supports four different types of customization needs going from
 **the most specific** to **the most generic**:
 
 1. [Specific variation for a one-time situation](#1-specific-variation-for-a-one-time-situation)
@@ -10,7 +10,7 @@ As components evolve in different contexts, Material-UI supports four different 
 
 ## 1. Specific variation for a one-time situation
 
-You might need to change the style of a component in some very specific situation, for which you have the following 
+You might need to change the style of a component in some very specific situation, for which you have the following
 solutions available:
 
 ### Overriding with class names
@@ -18,10 +18,10 @@ solutions available:
 The first way to override the style of a component is to use **class names**.
 Every component provides a `className` property which is always applied to the root element.
 
-In this example, we are using the [`withStyles()`](http://localhost:3000/customization/css-in-js#api) higher-order 
+In this example, we are using the [`withStyles()`](http://localhost:3000/customization/css-in-js#api) higher-order
 component to inject custom styles into the DOM, and to pass the class name to the `OverridesClasseNames` component via
- its `classes` prop. You can choose any other styling solution, or even plain CSS to create the styles, but be sure to 
-consider the [CSS injection order](/css-in-js#css-injection-order), as the CSS injected into the DOM by Material-UI to 
+ its `classes` prop. You can choose any other styling solution, or even plain CSS to create the styles, but be sure to
+consider the [CSS injection order](/css-in-js#css-injection-order), as the CSS injected into the DOM by Material-UI to
 style a component has the highest specificity possible since the `<link />` is injected at the bottom of the `<head />`
 to ensure the components always render correctly.
 
@@ -29,14 +29,14 @@ to ensure the components always render correctly.
 
 ### Overriding with classes
 
-When the `className` property isn't enough, and you need to access deeper elements, you can take advantage of the 
-`classes` property to customize all the CSS injected by Material-UI for a given component. The list of classes for each 
+When the `className` property isn't enough, and you need to access deeper elements, you can take advantage of the
+`classes` property to customize all the CSS injected by Material-UI for a given component. The list of classes for each
 component is documented in the **Component API** section. For instance, you can have a look at the [Button CSS API](/api/button#css-api).
 Alternatively, you can always look at the [implementation details](https://github.com/callemall/material-ui/blob/v1-beta/src/Button/Button.js).
 
 This example also uses `withStyles()` (see above), but `OverridesClasses` is using Button's `classes` prop to provide an
-array containing the **names of classes to override** (keys), and the **CSS class names to apply** (values). The 
-component's existing classes will continue to be injected, so it is only necessary to provide the specific styles you 
+array containing the **names of classes to override** (keys), and the **CSS class names to apply** (values). The
+component's existing classes will continue to be injected, so it is only necessary to provide the specific styles you
 wish to add or override.
 
 Notice that in addition to the button styling, the button label's capitalization has been changed:
@@ -55,32 +55,32 @@ You don't have to worry about CSS specificity as the inline-style takes preceden
 
 ## 2. Specific variation of a component
 
-You might need to create a variation of a component and use it in different contexts, for instance a colorful button on 
+You might need to create a variation of a component and use it in different contexts, for instance a colorful button on
 your product page, however you probably want to keep your code [*DRY*](https://en.wikipedia.org/wiki/Don%27t_repeat_yourself).
 
-The best approach is to follow option 1 and then take advantage of the composition power of React by exporting your 
+The best approach is to follow option 1 and then take advantage of the composition power of React by exporting your
 customized component to use wherever you need it.
 
 {{demo='pages/customization/OverridesComponent.js'}}
 
 ## 3. Material Design variations
 
-The Material Design specification documents different variations of certain components, such as how buttons come in 
-different shapes: 
-[flat](https://material.io/guidelines/components/buttons.html#buttons-flat-buttons), 
-[raised](https://material.io/guidelines/components/buttons.html#buttons-raised-buttons), 
+The Material Design specification documents different variations of certain components, such as how buttons come in
+different shapes:
+[flat](https://material.io/guidelines/components/buttons.html#buttons-flat-buttons),
+[raised](https://material.io/guidelines/components/buttons.html#buttons-raised-buttons),
 [floating](https://material.io/guidelines/components/buttons-floating-action-button.html) and more.
 
-Material-UI attempts to implement all of these variations. Please refer to the [Supported Components](/getting-started/supported-components) 
+Material-UI attempts to implement all of these variations. Please refer to the [Supported Components](/getting-started/supported-components)
 documentation to find out the current status of all supported Material Design components.
 
 ## 4. User global theme variation
 
-In order to promote consistency and manage the user interface as a whole, Material-UI provide mechanisms to apply global 
+In order to promote consistency and manage the user interface as a whole, Material-UI provide mechanisms to apply global
 changes. You can tweak the [configuration variables](/customization/themes#configuration-variables).
 
 ### Customizing all instances of a component type
 
-When the configuration variables aren't powerful enough, you can take advantage of the `overrides` key of the `theme` 
-to potentially change every single style injected by Material-UI into the DOM. Learn more about it in 
+When the configuration variables aren't powerful enough, you can take advantage of the `overrides` key of the `theme`
+to potentially change every single style injected by Material-UI into the DOM. Learn more about it in
 [the theme section of the documentation](/customization/themes#customizing-all-instances-of-a-component-type).
