+++ docs/src/pages/getting-started/example-projects.md
@@ -1,4 +1,4 @@
-# Examples
+# Example Projects
 
 Are you looking for an example project to get started?
