+++ pages/getting-started/example-projects.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import withRoot from 'docs/src/modules/components/withRoot';
 import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
-import markdown from 'docs/src/pages/customization/api.md';
+import markdown from 'docs/src/pages/getting-started/example-projects.md';
 
 function Page() {
   return <MarkdownDocs markdown={markdown} />;
