+++ pages/guides/api.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import withRoot from 'docs/src/modules/components/withRoot';
 import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
-import markdown from 'docs/src/pages/getting-started/examples.md';
+import markdown from 'docs/src/pages/guides/api.md';
 
 function Page() {
   return <MarkdownDocs markdown={markdown} />;
